feat(ref): replace wrappers with custom scripts for Ensembl downloads
The commit replaces the use of external wrappers for downloading Ensembl genome and annotation data with custom Bash scripts. This change provides more control over the download process, allows for better error handling, and integrates the use of `lftp` for parallel downloads. The `create_dict` rule has been commented out as it is no longer needed in the current workflow.